refactor: 시도 횟수만큼 자동차 움직이는 로직을 Game이 담당하도록
