+++ src/main/java/racingcar/game/Game.java
@@ -7,6 +7,7 @@ import racingcar.rule.NameDelimiter;
 import racingcar.view.InputView;
 import racingcar.view.OutputView;
 
+import java.util.ArrayList;
 import java.util.List;
 
 public class Game {
@@ -25,7 +26,7 @@ public class Game {
     public void play() {
         final RacingCars racingCars = createRacingCars();
         final MoveCount moveCount = createMoveCount();
-        final List<List<CarDto>> result = racingCars.startWith(moveCount);
+        final List<List<CarDto>> result = startRaceWith(racingCars, moveCount);
         outputView.printResult(result);
         final WinnersDto winners = racingCars.getWinners();
         outputView.printWinners(winners);
@@ -51,5 +52,14 @@ public class Game {
             outputView.printError(e.getMessage());
             return createMoveCount();
         }
+    }
+
+    List<List<CarDto>> startRaceWith(RacingCars racingCars, MoveCount moveCount) {
+        final List<List<CarDto>> result = new ArrayList<>();
+        for (int i = 0; i < moveCount.get(); i++) {
+            final List<CarDto> moveCar = racingCars.move();
+            result.add(moveCar);
+        }
+        return result;
     }
 }
